Improve CDF reading speed by scanning in batch
This PR improves CDF reading speed and reduces the number of tasks by creating a single scan for all CDF files from all versions(*). Compared to the previous approach (which creates a scan for each version), the new method significantly reduces the number of cloud requests and puts less stress on the Spark scheduler.

(*) Due to a limitation in how we broadcast DVs, the combining will happen for files that have only one associated DV across the entire requested version range.

GitOrigin-RevId: faf9a5ce062914ee182ff86d5f5696aeb62d1421