[StorageS3DynamoDB] Minor exception message changes
- [ ] Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [X] Other (storage-s3-dynamodb)

I was trying to track down the source of a weird FileSystemException and eventually discovered what it meant after looking at the source code. This is a minor change to the language used to help push you into a direction to discover the actual issue. My original change was to introduce a new, more specific exception type, but that would likely be a more painful breaking change.

Executed unit tests locally.

Yes, exception message text changed which if being explicitly tested against may fail.

Closes delta-io/delta#1972

Signed-off-by: Venki Korukanti <venki.korukanti@databricks.com>
GitOrigin-RevId: 8f18d5f8580ff9234cb2b4374080af1fcfd1bc71