Fix to Delta Uniform to support convert Delta null partition values to iceberg
The existing Delta to iceberg conversion has a bug that it does not handle null partition values as it will write the string with content "null" in the partition path, and "null" cannot be converted to other numeric types. The fix uses a special marker from iceberg library so it recognizes the null value and converts correctly.

GitOrigin-RevId: 667e795ead753803565340abcc23ae01d9738a2c