Rewrite DescribeDeltaDetailCommand to v1/v2 hybrid
This PR updates the Delta Lake DESCRIBE DETAIL command to work with V1 and V2 tables.

It rewrites the Delta DESCRIBE DETAIL command to use Spark's table resolution logic instead of resolving the target table manually at command execution time. For that, it changes `DescribeDeltaDetailCommand` from a `LeafRunnableCommand` to a `UnaryLike` command that takes an `UnresolvedTable` as a child plan node, which will be resolved by Spark. In addition, it also uses an `UnresolvedPath` LogicalPlan for the case that DESCRIBE DETAIL is run against a raw path. This logical plan is an indication that the target table is specified as a path so no table resolution is necessary.

Closes delta-io/delta#2012

GitOrigin-RevId: 33e35c225f0da0f970578e3ca065eaf7cb57c9bc