[Spark] Rename DeltaTableV2.snapshot as initialSnapshot
Closes https://github.com/delta-io/delta/pull/2063

`DeltaTableV2.snapshot` can be stale, if the object was instantiated long before the command using it. This can happen, for example, with a `DeltaTable` (which binds an internal `DeltaTableV2` when constructed and never refreshes it). This can be dangerous, so we rename the field and document the danger so it is less likely to cause bugs.

GitOrigin-RevId: 76e69f6889823dba8183cbd8f5e8d2671100d500