Allow user to create table with varchar type on a pre-existing location
When user run `create table (a varchar(100)) using delta location xx` twice with the same delta location, an error message reports the new schema does not match the existing one, while in fact they match.

Root cause is that Spark convert `varchar` columns as `string` + a key-value pair `["__CHAR_VARCHAR_TYPE_STRING" -> "varchar(<length>)"]` in metadata. But when Delta compares new schema with the existing one, it does not convert. Comparing `varchar` vs. `string` causes a mismatch and the error.

We apply the conversion to the new schema before comparing it against the existing schema to fix the problem.

GitOrigin-RevId: 645bdd48a8dd6b8c0ac121551c27d91c8f81cf68