Protocol version downgrade support in Delta Tables
Currently, we support Protocol downgrade in the form of feature removal but downgrading protocol versions is not possible. This PR adds support for protocol version downgrade. This is only allowed for tables that support either reader+writer table features or writer table features. The protocol downgrade takes place when the user removes a table feature and there are no non-legacy table features left in the table. The protocol is downgraded to the minimum reader/writer versions required to support all enabled legacy features.

For example, `Protocol(3, 7, readerFeatures=(DeletionVectors), writerFeatures=(DeletionVectors, ChangeDataFeed)` is downgraded to `Protocol(1, 4)` after removing the DeletionVectors table feature.

Closes delta-io/delta#2061

GitOrigin-RevId: 76633f6a08ae747ea508ef84e4e4f62a7ad5609d