[Spark] Minor update to `generate_iceberg_jars.py` to be specific about exclude files
The script `generate_iceberg_jars.py` breaks the build if the delta source is cloned in a directory with the name `source` or `opensource`. Update the script to be specific about the files to exclude (it should be `sources.jar`).

Ran `build/sbt clean publishM2` successfully after deleting the `icebergShaded/libs` folder and was able to see the generated libs.

Closes delta-io/delta#2121

Signed-off-by: Venki Korukanti <venki.korukanti@databricks.com>
GitOrigin-RevId: dbc0f87b3054c06c57eec699dba9349151260de1