[Spark] Replaces startTransaction calls with catalogTable overload
This PR replaces all calls to OptimisticTransaction.startTransaction() with calls to OptimisticTransaction.startTransaction(Option[CatalogTable]). This PR is part of https://github.com/delta-io/delta/issues/2105 and ensures that transactions have a valid catalogTable attached to them so Uniform can correctly update the table in the catalog.

This is a small refactoring change so existing test coverage is sufficient.

No

Closes https://github.com/delta-io/delta/pull/2125

GitOrigin-RevId: 014a459275e5fec5bfc51ff143563b2949c607c3