[Spark] Replaces startTransaction(Snapshot) calls with catalogTable overload
This PR replaces all calls to DeltaLog.startTransaction(Snapshot) with calls to DeltaLog.startTransaction(Option[CatalogTable], Snapshot). This PR is part of https://github.com/delta-io/delta/issues/2105 and a follow-up to https://github.com/delta-io/delta/pull/2125. It makes sure that transactions have a valid catalogTable attached to them so Uniform can correctly update the table in the catalog.

This is a small refactoring change so existing test coverage is sufficient.

No

Closes https://github.com/delta-io/delta/pull/2126

GitOrigin-RevId: d82787c64979a2dd4a363bf92a1640b7635ec02e