Reduce DeltaRetentionSuite flakiness by making ManualClock initialization consistent
Most RetentionSuite tests rely on ManualClock to time travel and trigger metadata cleanup. Cleanup boundaries are determined by finding files that were modified before 00:00 of the day on which currentTime-LOG_RETENTION_PERIOD falls. This means that for a long running test started at 23:59, the number of expired files would jump suddenly in 1 minute (the expiration boundary would move by a day as soon as system clock hits 00:00 of the next day). By fixing the start time of the test to 01:00, we avoid these scenarios. This should reduce test flakiness.

GitOrigin-RevId: 9540daec4f65174fb6a089b31011a70f57ecb770