[Kernel] Minor test clean up
(Cherry-pick of c7932974981ccc45c81ba279bef08bd12a0b18c3 to branch-3.0)

Remove the golden table `dv-with-columnmapping` from the `kernel` module and add it the `golden-tables` module (generated using `GoldenTables.scala`. This was done originally because of delta-io/delta#1886 which was fixed in the current master.

Ran the `DeletionVectorSuite` and it passes.