[Spark] Allow V2 checkpoints with no sidecars
The current code implicitly requires that all V2 checkpoints should have sidecars even though the Delta specification does not require this. This PR relaxes this assumption. Adds a new test which ensures that a snapshot can be initialized from a V2 Checkpoint where all of the actions are in the manifest and no sidecars are present.

GitOrigin-RevId: 1c1dc158898dd1c66c999ed4d1b52d51380fd23e