Fix DomainMetadata updates for DataFrameWriterV1 overwrite mode.
Before this change, it was not handling DomainMetadata updates (METADATA_DOMAINS_TO_REMOVE_FOR_REPLACE_TABLE) for DataFrameWriterV1 overwrite mode correctly. This change adds the handling in WriteIntoDelta.scala to only remove the DomainMetadata in METADATA_DOMAINS_TO_REMOVE_FOR_REPLACE_TABLE when it is an overwrite mode and overwriteSchema is true.

GitOrigin-RevId: c21c2f0ff9802064a9fe17c53efc4cd79386ac91