Introduce ClusteringTableFeature and CLUSTERED_BY file tag
https://github.com/delta-io/delta/issues/1874 requests Liquid clustering, and this PR starts the first step to introduce ClusteringTableFeature and CLUSTERED_BY tags.

When creating a clustered table, The feature clustering must exist in the table protocol's writerFeatures.

When a clustering implementation clusters files, writers must incorporate a tag with CLUSTERED_BY as the key and the name of the clustering implementation as the corresponding value in add action.

More detail can be found in the Delta protocol change PR https://github.com/delta-io/delta/pull/2264

The next step is to pave the way to integrate the table feature and clusterby tags when defining and clustering a clustered table.
Closes delta-io/delta#2281

GitOrigin-RevId: e210b491a324a0794ec9f3a9236bb1932a6677e3