Fix a delta protocol bug for DomainMetadata action.
DomainMetadata is listed as a write only feature, but it says
```
Readers must preserve all domains even if they don't understand them, i.e. the snapshot read must include them.
```
This doesn't seem right for a writer-only feature as if the reader doesn't understand metadata domains at all, the snapshot wouldn't include that column. This PR is fixing the problem.

Closes delta-io/delta#2288

GitOrigin-RevId: 911c1af6dd0347d88702616d5e19545ab5b2f305