Replace IndexedFile.isLast by END_INDEX to reliably move to BASE_INDEX of next version
Remove `IndexedFile.isLast` and replace it by a `DeltaSourceOffset.END_INDEX` that is always included after the last entry.

There were several problems with `isLast`. For one thing, we were not setting it for the last file in initial snapshots, so when a streaming query starts and processes the initial snapshot, it would stop at the last file in the snapshot. Then on the next trigger we would reconstruct the initial snapshot only to find out that we'd already processed all of it.

Another problem with `isLast` is that it doesn't exist for empty commits or snapshots. As a result, the stream would not be able to move forward if it started at the `BASE_INDEX` of an empty commit.

With this change, I added `END_INDEX` to the end of every snapshot (CDC and non-CDC), and also a BASE_INDEX to the start (even if it's technically not used, but it's good to be consistent).

I also updated some tests in `DeltaCDCStreamSuite` that were incorrect. They restarted a query without a checkpoint, and then did checks assuming that the query had restarted from a checkpoint. I also changed some tests to be very specific about the offsets that they expect the batches in the stream to end at.
Closes delta-io/delta#2241

GitOrigin-RevId: 25eb321c2c85d0db9c89e0d0caf8aa6b3febd5a7