[Spark] Use IndexedFile for DeltaSource admission control instead of FileAction
This is a small refactoring. The DeltaSource has admission control (e.g. maxFilesPerTrigger), and the code is based on `FileAction` objects. However, all the calling code really has `IndexedFile` objects, and has to do manual conversions to deal with the fact that not all `IndexedFile` objects have file actions. This PR simplifies this by making the admission control based on `IndexedFile`.

Closes delta.io/delta#2317

GitOrigin-RevId: 683e6369cc5383a7863f15bdbdc87c4849b36d45