Make DeltaSourceOffset versionless
The `DeltaSourceOffset` still has a version, but it's supposed to always be `VERSION_3`. This turned out to not always be the case. Furthermore, it was not guaranteed that all use cases would use the new `BASE_INDEX`, as some of them still used a hardcoded `-1`. This PR removes the concept of a version from `DeltaSourceOffset` and makes it entirely versionless. The versioning is retained in the serialized form where it belongs.

Closes delta-io/delta#2335

GitOrigin-RevId: 835f15d3adf83a47b547344b6087626c93c26222