[Spark] DeltaAnalysis should not turn the target table of MERGE into v1 relation too early
In the Spark analyzer, it skips resolving the MERGE command if the target table is v2 and reports `ACCEPT_ANY_SCHEMA`. However, the `DeltaAnalysis` will turn Delta v2 relation to v1, and make Spark mistakenly resolve the Delta MERGE commands, which can lead to issues as Delta supports more features like schema evolution.

We can't simply move the "turn Delta v2 relation to v1" code to the post-resolution phase, as we still need to do it to resolve metadata columns, which heavily rely on v1 relation. This PR does a surgical fix: if we find a MERGE command that can't be turned into its Delta variant yet (because the source query is not resolved), we mark its target table and do not turn it into v2 relation. After the MERGE command is fully resolved, we will turn it to the Delta variable, and all Delta relation inside it will become v1 relations

Closes delta-io/delta#2308

GitOrigin-RevId: 6548dea6889215e04c7a1513b6028d42f651b2eb