[Spark] Create MergeCDCCoreSuite to run only core CDC tests
#### Which Delta project/connector is this regarding?

- [X] Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description
This is a small refactor to `MergeCDCSuite` to create a second, smaller suite `MergeCDCCoreSuite` that only runs CDC tests defined in that same file.
This allows running a small number of tests with high CDC coverage instead of running the (very) large number of tests defined in `MergeIntoSQLSuite` every time.

Running the tests

## Does this PR introduce _any_ user-facing changes?
No

Closes delta-io/delta#2385

GitOrigin-RevId: ceed6022e75ade4ea40359876e85c2453b624c6c