[Spark] Adds code to support Auto Compaction.
Auto compaction combines small files within partitions to reduce problems due to a proliferation of small files. Auto compaction is implemented as a post commit hook, and so occurs after the write to a table has succeeded. It runs synchronously on the cluster that has performed the write.

You can control the output file size by setting the  spark.databricks.delta.autoCompact.maxFileSize.

Auto compaction is only triggered for partitions or tables that have at least a certain number of small files. You can optionally change the minimum number of files required to trigger auto compaction by setting spark.databricks.delta.autoCompact.minNumFiles.

This PR creates a post commit hook, which runs an `OptimizeExecutor` (from `OptimizeTableCommand.scala`), which will do the compaction.

We add a post-commit hook in TransactionalWrite, that will check if auto-compaction is needed. If the configs are set such that the write meets the criteria (i.e. AC is enabled, enough small files exist, etc) then partitions that meet the criteria will be reserved, and used to make an OptimizeExecutor targeting those partitions, and with the appropriate config values.

This runs and will compact the files. Partitions are then released for future compactions to consider.

AutoCompact is *disabled* by default

There are a number of new configs introduced by this PR, all with prefix spark.databricks.delta.autoCompact.  Through a lot of experimentation and user feedback, we found these values to work well across a large range of tables and configurations.

- `autoCompact.enabled`: should auto compaction run? (default: false)
- `autoCompact.maxFileSize`: Target file size produced by auto compaction (default: 128 MB)
- `autoCompact.minFileSize`: Files which are smaller than this threshold (in bytes) will be grouped together and rewritten as larger files by the Auto Compaction. (default: half of maxFileSize)

Closes delta-io/delta#2414

GitOrigin-RevId: b0fb8f09a5c13d2ac14a907e9e872646354e7539