[Spark] Widen non-deterministic predicates during partition-level conflict checking
During conflict checking we use the partition predicates of the transaction to see if any newly added files of the winning transaction should have been read by the current transaction. This causes a re-evaluation of non-deterministic predicates. This can lead to not detecting conflicts, even though they should have been detected. To fix this, we widen non-deterministic predicates to `TrueLiteral`s, to always conflict with more files.

Closes delta-io/delta#2363

GitOrigin-RevId: 4745ce2f36caad724171a282e22da87e0fc68594