Avoid redudant jobs in buildRowIndexSetsForFilesMatchingCondition.
In `buildRowIndexSetsForFilesMatchingCondition` we join the results DataFrame with the candidate file list to fetch the Deletion Vector Descriptors. Currently, there is a sanity check there to verify we do not eliminate any rows due to path encoding issues. The sanity check compares before and after counts. The counts produce 2 extra jobs which can cause a significant performance overhead. For example, in the case of merge each job could be a full shuffle join between the source and the target. This PR eliminates the extra jobs while maintaining the sanity check.

Closes delta-io/delta#2516

GitOrigin-RevId: e9f1c17c482be35979a64b2ee5e72c4562c084d5