[Spark][3.1] Emphasize paths should (or should not) be escaped in DV helper
This PR is a backport of #2510 to the 3.1 branch to simplify further cherry-picks to branch-3.1 in future.

This PR renames stringToPath and pathToString methods, this will emphasize the input/output string should/will be escaped.

We also added a third new method unescapedStringToPath. which is used by almost all calls from tests.