[Spark] Improve exception in resolveQueryColumnsByName
In this PR, we replace an assertion error with an actual exception. This improves the error message from an internal error in spark to a more actionable error indicating user error. The assertion comes from an assumption that the code in DeltaAnalysis is run after PreprocessTableInsertion (resolution), but this is not the case. We choose to duplicate the check instead of reordering the rules to have a smaller change.
Closes delta-io/delta#2555

GitOrigin-RevId: 5c61f919852c72b28a944abf46b44f5de3e61ed6