[Spark] Always materialize the merge source if it contains a UDF
#### Which Delta project/connector is this regarding?

-Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description

This PR fixes a correctness issue in merge due the source query returning different results in job 1 and job 2. This issue is present when the source contains a non-deterministic UDF that has been marked as deterministic. UDFs are often incorrectly marked as deterministic, and therefore we should not trust this information and instead always materialize the source if it contains a UDF

Added a test to `MergeIntoMaterializeSourceSuite`.

## Does this PR introduce _any_ user-facing changes?

No

Closes delta-io/delta#2585

GitOrigin-RevId: 1b55acc398d8c8e36dd618ebba28c92ed3309222