Normalize column names in the nested fields when writing
Function `SchemaUtils.normalizeSchemaNames()` is used during a Delta write. It was intended to correct the case of names of any top level-fields that differed between the input schema and the table. If the case of any nested field differed, the function was supposed to raise an exception.

However, due to a long-standing bug, the function could instead ignore the difference in the nested fields. This results in a data corruption. While the column values are written into the Delta table, the stats for these column are not gathered correctly. Instead, the stats are recorded as-if the column was missing in the input, that is: `minValue = null`, `maxValue = null`, `nullCount = rowCount`.

This PR implements the full normalization of nested field names to fix this.

Closes delta-io/delta#2569

GitOrigin-RevId: 8139d9c883b053d0ce4e27dd9a2db3214df1b5b5