Factor logic to collect files to REORG out of OPTIMIZE
#### Which Delta project/connector is this regarding?

-Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description

This is a plain refactor of REORG TABLE / OPTIMIZE to allow for better extendability and adding new types of REORG TABLE operations in the future.

The REORG operation currently supports:
- PURGE: remove soft deleted rows (DVs) and dropped columns.
- UPGRADE UNIFORM: rewrite files to be iceberg compatible.

More operations can be used in the future to allow dropping table features, in particular for column mapping: rewrite files to have the physical column names match the logical column name.

This a plain refactor without functional changes.
Closes delta-io/delta#2616

GitOrigin-RevId: b8e8ad4d148201a33b1fb173ebcfe4ad8b8407ef