[Delta Uniform] overwrite source column field id for Iceberg PartitionField with field id assigned by Delta
context: Delta and Iceberg traverse schema and assigns field id in different way. Delta uniform currently use a extra Iceberg txn to overwrite the schema in iceberg table with wrong field ids reassigned by Iceberg.

However, if the source field id for partition columns is different in that schema overwrite txn, Iceberg always expect field id for partition columns to be the same and does not have logic to reconcile that, and then Iceberg will fail the overwrite txn. This PR adds the logic to adopt new source column field id to Iceberg PartitionField if changed, so the overwrite txn can go through and set the correct source column field id for PartitionFields.

Closes delta-io/delta#2676

GitOrigin-RevId: 37be472e9794d0a87c59a8fe06efe237ee1c609e