Introduce InCommitTimestamp feature and write monotonically increasing timestamps in CommitInfo
Follow-up for https://github.com/delta-io/delta/issues/2532.

Adds a new writer feature called `inCommitTimestamp`. When this feature is enabled, the writer will make sure that it writes `commitTimestamp` in CommitInfo which contains a monotonically increasing timestamp.

This PR is an initial implementation, it does not handle timestamp retrieval efficiently. It does not try to populate the inCommitTimestamp in Snapshot even in places where it is already available, instead Snapshot has to perform an IO to read the timestamp.

Closes delta-io/delta#2596

GitOrigin-RevId: 44904e734eee74378ee55f708beb29a484cd93e6