[Spark] Allow specifying custom prefixes for temp folders used by tests
#### Which Delta project/connector is this regarding?

- [X] Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description

This PR introduces a new trait, `DeltaSQLTestUtils`, to be used by all tests as a replacement for Apache Spark's `SQLTestUtils.` The new trait provides an ability for each test to specify a custom name prefix for temporary folders, so we can test the behavior of Delta Lake when a table's path contains special characters.

Not needed.

## Does this PR introduce _any_ user-facing changes?

No.

Closes delta-io/delta#2696

Signed-off-by: Paddy Xu <xupaddy@gmail.com>
GitOrigin-RevId: fa9bead06ff6da1d7baff58a564fbe4cbeac6727