[Spark] In materialize merge source: Only log storage error if we actually materialized (#2727)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-io/delta/blob/master/CONTRIBUTING.md
2. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  3. Be sure to keep the PR description updated to reflect all changes.
  4. Please write your PR title to summarize what this PR proposes.
5. If possible, provide a concise example to reproduce the issue for a
faster review.
6. If applicable, include the corresponding issue number in the PR title
and link it in the body.
-->

#### Which Delta project/connector is this regarding?
<!--
Please add the component selected below to the beginning of the pull
request title
For example: [Spark] Title of my pull request
-->

- [x] Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description

Be more defensive about calling `Option.get` on `materializedSourceRDD`
and check first that it's actually `Some`. In fact only invoke the
entire branch where we log only when we actually did materialize the
source (and not, say, ran out of disk space before).

## How was this patch tested?


This PR adds a new test for the scenario where we throw an out of disk
space error without having materialized the source.

## Does this PR introduce _any_ user-facing changes?

No